LayoutTests/fast/parser/fragment-parser-expected.txt

Test that fragment parsing does not affect the host document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parse('<span><body bgcolor=red>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><html bgcolor=red>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><meta>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><base>') is ['<span></span>','<html><body></body></html>']
PASS parse('<html><script>') is ['','<html><body></body></html>']
PASS parse('<html><style>') is ['','<html><body></body></html>']
PASS parse('<html><meta>') is ['','<html><body></body></html>']
PASS parse('<html><link>') is ['','<html><body></body></html>']
PASS parse('<html><object>') is ['<object></object>','<html><body></body></html>']
PASS parse('<html><embed>') is ['<embed>','<html><body></body></html>']
PASS parse('<html><title>') is ['','<html><body></body></html>']
PASS parse('<html><isindex>') is ['<div><hr>This is a searchable index. Enter search keywords: <isindex type="khtml_isindex"><hr></div>','<html><body></body></html>']
PASS parse('<html><base>') is ['','<html><body></body></html>']
PASS parse('<html><div>') is ['<div></div>','<html><body></body></html>']
PASS parse('<frameset>') is ['<frameset></frameset>','<html><body></body></html>']
PASS parse('<html>x', true) is ['x','no document element']
PASS successfullyParsed is true

TEST COMPLETE
